backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
groq>=0.12.0
python-dotenv==1.0.1
pydantic==2.9.0
websockets==13.0
python-multipart==0.0.9
